+++ src/cce-inventory-item/cce-inventory-item.service.js
@@ -43,13 +43,18 @@
             },
             update: {
                 method: 'PUT'
+            },
+            transfer: {
+                method: 'PUT',
+                url: cceUrlFactory('/api/inventoryItems/:id/transfer')
             }
         });
 
         return {
             get: get,
             query: query,
-            save: save
+            save: save,
+            transfer: transfer
         };
 
         /**
@@ -99,6 +104,26 @@
         function save(inventoryItem) {
             if (inventoryItem.id) {
                 return resource.update({
+                    id: inventoryItem.id
+                }, inventoryItem).$promise;
+            }
+            return resource.save({}, inventoryItem).$promise;
+        }
+
+        /**
+         * @ngdoc method
+         * @methodOf cce-inventory-item.inventoryItemService
+         * @name transfer
+         *
+         * @description
+         * Changes the Inventory facility, program and istallation date
+         *
+         * @param  {Object}     inventoryItem   the updated inventory item
+         * @return {Promise}                    the promise resolving to the updated item
+         */
+        function transfer(inventoryItem) {
+            if (inventoryItem.id) {
+                return resource.transfer({
                     id: inventoryItem.id
                 }, inventoryItem).$promise;
             }